Major Changes
Removed the install api function as i found out that i didn't need it,
added more pieces of auth, server and client now support more commands,
now there is even server client comms, but also don't use swearing in
you username as we can now force-change user names. this is going to be
the first of sever server-client updates, commands can now be executed
on the client through the server. also the server now has fixed the
/server commands so now we can use them again (darned php can't decide
to use 0 or 1 as the beginning of explode) server also has 1 more
command the ccmd or client command.